requirements.txt
Flask==1.1.2
google-cloud-secret-manager==2.2.0
stripe>=2.48.0,<3.0
gunicorn
